fix: fix theme and bill's history, add user credit rate, fix p2p and me2me transactions

+++ client_Customer/app/src/main/java/com/example/customerclient/domain/repositories/TransactionRepository.kt
@@ -17,7 +17,7 @@ interface TransactionRepository {
         message: String?
     )
 
-    suspend fun deposit(billId: String, amount: Int)
+    suspend fun deposit(billId: String, amount: Double)
 
-    suspend fun withdraw(billId: String, amount: Int)
+    suspend fun withdraw(billId: String, amount: Double)
 }
